Update cover.py
Changed open and close levels. With current positions it is imoossible to use integration with motorized windows. For example I want to slightly open window during the winter to let some fresh air in, but don' freeze to death :) Then I open window at as low as 3% and it's enough. With current level it will be shown as fully closed in the interface which is not right.